Enum growable::Growable

pub enum Growable {
    Some {
        len: usize,
        ptr_alignment: usize,
        ptr: Unique<u8>,
    },
    None,
}

A chunk of heap memory that can be assigned to a struct or a trait object. Until assigned to some data it behaves similarly to a Box<[u8; N]>, it can be cloned and would be dropped if leaves the scope.

Examples

First, let's spawn a new Growable. In this case no allocation would be performed on init.

  let growable = Growable::new();

Now we can assign some data to it.

  let arr: Reusable<[char; 3]> = growable.assign(['f', 'o', 'o']);
  assert_eq!(&*arr, &['f', 'o', 'o']);

Unwanted data could be then freed on demand, fetching Growable back. Then it could be assigned to some data again and so on.

  let growable = arr.free();
  let arr: Reusable<[char; 6]> = growable.assign(['f', 'o', 'o', 'b', 'a', 'r']);
  assert_eq!(&*arr, &['f', 'o', 'o', 'b', 'a', 'r']);

Variants
Some

Pre-allocated chunk of memory.

Fields of Some

len: usize

Memory block length.

ptr_alignment: usize

Required alignment for the pointer.

ptr: Unique<u8>

Pointer.

None

No assigned memory.

Methods

impl Growable

pub fn new() -> Self

Returns a new instance of Growable but does not allocate any memory on the heap yet.

pub fn with_capacity_for<T>() -> Self

Returns a new instance of Growable with memory already allocated on the heap suitable to store an instance of a given type T.

pub fn with_capacity(len: usize, ptr_alignment: usize) -> Self

Returns a new instance of Growable with memory already allocated on the heap. In stable Rust it is possible to get a required pointer alignment for any type with align_of function.

pub fn len(&self) -> usize

Returns the amount of memory allocated by this Growable.

pub fn is_empty(&self) -> bool

Returns true if no memory has been allocated yet.

pub fn ptr_alignment(&self) -> Option<usize>

Returns an alignment of an underlying memory pointer if any.

pub fn assign<T>(self, t: T) -> Reusable<T> where
T: 'static,

Returns allocated on the heap struct, an actual (re)allocation will be performed only if there is not enough space or the pointer alignment is invalid.

pub fn assign_as_trait<T, U: ?Sized>(self, t: T) -> Reusable<U> where
T: Unsize<U> + 'static,
U: 'static,

Returns allocated on the heap struct, an actual (re)allocation will be performed only if there is not enough space or the pointer alignment is invalid. Additionally stores a pointer to the vtable creating trait object.

Examples

  trait Answerer {

      fn get_answer(&self) -> u32;
  }

  struct Foo;

  impl Answerer for Foo {

      fn get_answer(&self) -> u32 { 42 }
  }

  let growable = Growable::new();
  let foo: Reusable<Answerer> = growable.assign_as_trait(Foo);
  assert_eq!(foo.get_answer(), 42);