Struct growable::Reusable

pub struct Reusable<T: ?Sized> { /* fields omitted */ }

Growable with some data assigned to it. It behaves just like default Box does (so it WILL free memory on drop) but also could be freed manually, fetching Growable back.

Methods

impl<T: ?Sized> Reusable<T> where
T: ,

pub fn free(self) -> Growable

Performs drop call on the stored value and returns freed memory block back as a Growable struct.