Crate growable

A growable, reusable box for Rust.

This crate provides a custom Box type with matching API that also allows to reuse the same box to store different types with minimal amount of allocations and is supposed to be used with a custom, pool-based allocator of user's choice such as Lifeguard, but also can be used as a standalone library.

Note that this crate uses a lot of ground-breaking features of Rust and therefore is only available on current Nightly build.

Structs

Reusable
Growable with some data assigned to it. It behaves just like default Box does (so it WILL free memory on drop) but also could be freed manually, fetching Growable back.

Enums

Growable
A chunk of heap memory that can be assigned to a struct or a trait object. Until assigned to some data it behaves similarly to a Box<[u8; N]>, it can be cloned and would be dropped if leaves the scope.